requirements.txt
alembic==1.7.5
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
bqplot==0.11.0
bs4==0.0.1
bson==0.5.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-conf==0.0.1
click-stream==0.0.10
cssselect==1.1.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
docopt==0.6.2
dominate==2.6.0
elasticsearch==7.16.0
elasticsearch-dsl==7.4.0
elasticsearch-loader==0.6.0
entrypoints==0.3
fake-useragent==0.1.11
feedparser==6.0.8
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
install==1.3.5
ipykernel==6.4.2
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.6.4
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mongo==0.2.0
multitasking==0.0.10
nbclient==0.5.9
nbconvert==5.6.1
nbformat==4.4.0
nbinteract==0.2.6
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyaml==21.10.1
pycparser==2.21
pyee==8.2.2
Pygments==2.10.0
pymongo==4.0
pyparsing==3.0.6
pyppeteer==0.2.6
pyquery==1.4.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.26.0
requests-html==0.10.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
shell==1.0.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
traittypes==0.2.1
urllib3==1.26.7
visitor==0.1.3
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
WTForms==3.0.0
yahoo-fin==0.8.9.1
yfinance==0.1.67
zipp==3.6.0
